+++ src/tg_bot_float_csm_wiki_source/services/csm_wiki_skin_data_dto.py
@@ -2,6 +2,7 @@ from typing import List
 
 from pydantic import BaseModel
 
+
 class CSMWikiSkinDataDTO(BaseModel):
     qualities: List[str] = []
     stattrak_existence: bool = False
